fix compile errors

+++ src/Shared/Logistics.Models/GrossesChart.cs
@@ -18,6 +18,6 @@ public class GrossesChart<T> where T : IGrossChart
         }
     }
 
-    public double TotalGross { get; private set; }
+    public decimal TotalGross { get; private set; }
     public double TotalDistance { get; private set; }
 }
